requirements.txt
geopandas
matplotlib
fiona
gdal
ogr
osr